Fixed macOS build compatibility
- Made compiler flags cross-platform (MSVC vs Clang/GCC)
- Fixed signed/unsigned comparison in io_helpers.cpp
- Removed unused variable in serial_helpers.cpp
- Fixed char vs unsigned char comparison with START_BYTE in knight_serial_interface.cpp
- Changed range-for loop to use const reference in knight_protocol_parser.cpp

+++ src/utils/io_helpers.cpp
@@ -16,9 +16,9 @@ void clear_lines(int line_count)
 std::vector<std::string> splitString(std::string& input, char delimiter)
 {
     std::vector<std::string> tokens;
-    int tokenStart = 0;
+    size_t tokenStart = 0;
 
-    for (int i = 0; i <= input.size(); i++)
+    for (size_t i = 0; i <= input.size(); i++)
     {
         if (input[i] == delimiter || input[i] == '\0')
         {
